backend/requirements.txt
fastapi
uvicorn[standard]
requests
websocket-client
python-multipart
